+++ src/ios/PurchasesPlugin.swift
@@ -1,0 +1,10 @@
+//
+//  PurchasesPlugin.swift
+//  HelloCordova
+//
+//  Created by Andrés Boedo on 9/23/20.
+//
+// This file is left intentionally blank,
+// the code for the plugin lives in CDVPurchasesPlugin
+
+import Foundation