+++ src/plugin/plugin.ts
@@ -195,6 +195,21 @@ export interface PurchasesEntitlementInfos {
   readonly active: { [key: string]: PurchasesEntitlementInfo };
 }
 
+export interface PurchasesTransaction { 
+  /**
+   * RevenueCat Id associated to the transaction.
+   */
+  readonly revenueCatId: string;
+  /**
+   * Product Id associated with the transaction.
+   */
+  readonly productId: string;
+  /**
+   * Purchase date of the transaction in ISO 8601 format.
+   */
+  readonly purchaseDate: string;
+}
+
 export interface PurchaserInfo {
   /**
    * Entitlements attached to this purchaser info
@@ -208,6 +223,11 @@ export interface PurchaserInfo {
    * Set of purchased skus, active and inactive
    */
   readonly allPurchasedProductIdentifiers: [string];
+  /**
+   * Returns all the non-subscription purchases a user has made.
+   * The purchases are ordered by purchase date in ascending order.
+   */
+  readonly nonSubscriptionTransactions: [PurchasesTransaction];
   /**
    * The latest expiration date of all purchased skus
    */
@@ -522,6 +542,9 @@ class Purchases {
 
   /**
    * Add a dict of attribution information
+   *
+   * @deprecated Use the set<NetworkId> functions instead.
+   *
    * @param {object} data Attribution data from any of the attribution networks in Purchases.ATTRIBUTION_NETWORKS
    * @param {ATTRIBUTION_NETWORK} network Which network, see Purchases.ATTRIBUTION_NETWORK
    * @param {string?} networkUserId An optional unique id for identifying the user. Needs to be a string.
@@ -981,7 +1004,7 @@ class Purchases {
   /**
    * Subscriber attribute associated with the push token for the user
    *
-   * @param pushToken null will delete the subscriber attribute.
+   * @param pushToken Empty String or null will delete the subscriber attribute.
    */
   public static setPushToken(pushToken: string | null) {
     window.cordova.exec(
@@ -991,6 +1014,186 @@ class Purchases {
       "setPushToken",
       [pushToken]
     );
+  }
+  
+  /**
+   * Subscriber attribute associated with the adjustID for the user
+   *
+   * @param adjustID Empty String or null will delete the subscriber attribute.
+   */
+  public static setAdjustID(adjustID: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setAdjustID",
+      [adjustID]
+    )
+  }
+  
+  /**
+   * Subscriber attribute associated with the appsflyerID for the user
+   *
+   * @param appsflyerID Empty String or null will delete the subscriber attribute.
+   */
+  public static setAppsflyerID(appsflyerID: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setAppsflyerID",
+      [appsflyerID]
+    )
+  }
+  
+  /**
+   * Subscriber attribute associated with the fbAnonymousID for the user
+   *
+   * @param fbAnonymousID Empty String or null will delete the subscriber attribute.
+   */
+  public static setFBAnonymousID(fbAnonymousID: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setFBAnonymousID",
+      [fbAnonymousID]
+    )
+  }
+  
+  /**
+   * Subscriber attribute associated with the mparticleID for the user
+   *
+   * @param mparticleID Empty String or null will delete the subscriber attribute.
+   */
+  public static setMparticleID(mparticleID: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setMparticleID",
+      [mparticleID]
+    )
+  }
+  
+  /**
+   * Subscriber attribute associated with the onesignalID for the user
+   *
+   * @param onesignalID Empty String or null will delete the subscriber attribute.
+   */
+  public static setOnesignalID(onesignalID: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setOnesignalID",
+      [onesignalID]
+    )
+  }
+  
+  /**
+   * Subscriber attribute associated with the mediaSource for the user
+   *
+   * @param mediaSource Empty String or null will delete the subscriber attribute.
+   */
+  public static setMediaSource(mediaSource: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setMediaSource",
+      [mediaSource]
+    )
+  }
+  
+  /**
+   * Subscriber attribute associated with the campaign for the user
+   *
+   * @param campaign Empty String or null will delete the subscriber attribute.
+   */
+  public static setCampaign(campaign: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setCampaign",
+      [campaign]
+    )
+  }
+  
+  /**
+   * Subscriber attribute associated with the adGroup for the user
+   *
+   * @param adGroup Empty String or null will delete the subscriber attribute.
+   */
+  public static setAdGroup(adGroup: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setAdGroup",
+      [adGroup]
+    )
+  }
+  
+  /**
+   * Subscriber attribute associated with ad token for the user
+   *
+   * @param ad Empty String or null will delete the subscriber attribute.
+   */
+  public static setAd(ad: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setAd",
+      [ad]
+    )
+  }
+  
+  /**
+   * Subscriber attribute associated with the keyword for the user
+   *
+   * @param keyword Empty String or null will delete the subscriber attribute.
+   */
+  public static setKeyword(keyword: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setKeyword",
+      [keyword]
+    )
+  }
+  
+  /**
+   * Subscriber attribute associated with the creative for the user
+   *
+   * @param creative Empty String or null will delete the subscriber attribute.
+   */
+  public static setCreative(creative: string | null) { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "setCreative",
+      [creative]
+    )
+  }
+
+  /**
+   * Automatically collect subscriber attributes associated with the device identifiers
+   * $idfa, $idfv, $ip on iOS
+   * $gpsAdId, $androidId, $ip on Android
+   */
+  public static collectDeviceIdentifiers() { 
+    window.cordova.exec(
+      null,
+      null,
+      PLUGIN_NAME,
+      "collectDeviceIdentifiers",
+      []
+    )
   }
   
   /**