+++ www/plugin.d.ts
@@ -175,6 +175,20 @@ export interface PurchasesEntitlementInfos {
         [key: string]: PurchasesEntitlementInfo;
     };
 }
+export interface PurchasesTransaction {
+    /**
+     * RevenueCat Id associated to the transaction.
+     */
+    readonly revenueCatId: string;
+    /**
+     * Product Id associated with the transaction.
+     */
+    readonly productId: string;
+    /**
+     * Purchase date of the transaction in ISO 8601 format.
+     */
+    readonly purchaseDate: string;
+}
 export interface PurchaserInfo {
     /**
      * Entitlements attached to this purchaser info
@@ -188,6 +202,11 @@ export interface PurchaserInfo {
      * Set of purchased skus, active and inactive
      */
     readonly allPurchasedProductIdentifiers: [string];
+    /**
+     * Returns all the non-subscription purchases a user has made.
+     * The purchases are ordered by purchase date in ascending order.
+     */
+    readonly nonSubscriptionTransactions: [PurchasesTransaction];
     /**
      * The latest expiration date of all purchased skus
      */
@@ -466,6 +485,9 @@ declare class Purchases {
     static setAllowSharingStoreAccount(allowSharing: boolean): void;
     /**
      * Add a dict of attribution information
+     *
+     * @deprecated Use the set<NetworkId> functions instead.
+     *
      * @param {object} data Attribution data from any of the attribution networks in Purchases.ATTRIBUTION_NETWORKS
      * @param {ATTRIBUTION_NETWORK} network Which network, see Purchases.ATTRIBUTION_NETWORK
      * @param {string?} networkUserId An optional unique id for identifying the user. Needs to be a string.
@@ -684,9 +706,81 @@ declare class Purchases {
     /**
      * Subscriber attribute associated with the push token for the user
      *
-     * @param pushToken null will delete the subscriber attribute.
+     * @param pushToken Empty String or null will delete the subscriber attribute.
      */
     static setPushToken(pushToken: string | null): void;
+    /**
+     * Subscriber attribute associated with the adjustID for the user
+     *
+     * @param adjustID Empty String or null will delete the subscriber attribute.
+     */
+    static setAdjustID(adjustID: string | null): void;
+    /**
+     * Subscriber attribute associated with the appsflyerID for the user
+     *
+     * @param appsflyerID Empty String or null will delete the subscriber attribute.
+     */
+    static setAppsflyerID(appsflyerID: string | null): void;
+    /**
+     * Subscriber attribute associated with the fbAnonymousID for the user
+     *
+     * @param fbAnonymousID Empty String or null will delete the subscriber attribute.
+     */
+    static setFBAnonymousID(fbAnonymousID: string | null): void;
+    /**
+     * Subscriber attribute associated with the mparticleID for the user
+     *
+     * @param mparticleID Empty String or null will delete the subscriber attribute.
+     */
+    static setMparticleID(mparticleID: string | null): void;
+    /**
+     * Subscriber attribute associated with the onesignalID for the user
+     *
+     * @param onesignalID Empty String or null will delete the subscriber attribute.
+     */
+    static setOnesignalID(onesignalID: string | null): void;
+    /**
+     * Subscriber attribute associated with the mediaSource for the user
+     *
+     * @param mediaSource Empty String or null will delete the subscriber attribute.
+     */
+    static setMediaSource(mediaSource: string | null): void;
+    /**
+     * Subscriber attribute associated with the campaign for the user
+     *
+     * @param campaign Empty String or null will delete the subscriber attribute.
+     */
+    static setCampaign(campaign: string | null): void;
+    /**
+     * Subscriber attribute associated with the adGroup for the user
+     *
+     * @param adGroup Empty String or null will delete the subscriber attribute.
+     */
+    static setAdGroup(adGroup: string | null): void;
+    /**
+     * Subscriber attribute associated with ad token for the user
+     *
+     * @param ad Empty String or null will delete the subscriber attribute.
+     */
+    static setAd(ad: string | null): void;
+    /**
+     * Subscriber attribute associated with the keyword for the user
+     *
+     * @param keyword Empty String or null will delete the subscriber attribute.
+     */
+    static setKeyword(keyword: string | null): void;
+    /**
+     * Subscriber attribute associated with the creative for the user
+     *
+     * @param creative Empty String or null will delete the subscriber attribute.
+     */
+    static setCreative(creative: string | null): void;
+    /**
+     * Automatically collect subscriber attributes associated with the device identifiers
+     * $idfa, $idfv, $ip on iOS
+     * $gpsAdId, $androidId, $ip on Android
+     */
+    static collectDeviceIdentifiers(): void;
     /**
      * Set this property to your proxy URL before configuring Purchases *only* if you've received a proxy key value from your RevenueCat contact.
      * @param url Proxy URL as a string.